refactor : Image폴더로 이동
