+++ core/domain/src/main/java/com/withpeace/withpeace/core/domain/model/image/ImageFolder.kt
@@ -1,4 +1,4 @@
-package com.withpeace.withpeace.core.domain.model
+package com.withpeace.withpeace.core.domain.model.image
 
 data class ImageFolder(
     val folderName: String,
