+++ core/domain/src/main/java/com/withpeace/withpeace/core/domain/model/post/RegisterPost.kt
@@ -1,6 +1,6 @@
 package com.withpeace.withpeace.core.domain.model.post
 
-import com.withpeace.withpeace.core.domain.model.LimitedImages
+import com.withpeace.withpeace.core.domain.model.image.LimitedImages
 
 data class RegisterPost(
     val title: String,
