+++ core/domain/src/test/kotlin/com/withpeace/withpeace/core/domain/usecase/GetAlbumImagesUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.withpeace.withpeace.core.domain.usecase
 
 import com.google.common.truth.Truth.assertThat
-import com.withpeace.withpeace.core.domain.model.ImagePagingInfo
+import com.withpeace.withpeace.core.domain.model.image.ImagePagingInfo
 import com.withpeace.withpeace.core.domain.paging.ImagePagingSource
 import com.withpeace.withpeace.core.domain.repository.ImageRepository
 import io.mockk.mockk
