+++ core/domain/src/test/kotlin/com/withpeace/withpeace/core/domain/usecase/RegisterPostUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.withpeace.withpeace.core.domain.usecase
 
 import com.google.common.truth.Truth.assertThat
-import com.withpeace.withpeace.core.domain.model.LimitedImages
+import com.withpeace.withpeace.core.domain.model.image.LimitedImages
 import com.withpeace.withpeace.core.domain.model.WithPeaceError
 import com.withpeace.withpeace.core.domain.model.post.PostTopic
 import com.withpeace.withpeace.core.domain.model.post.RegisterPost
