+++ feature/registerpost/src/test/java/com/withpeace/withpeace/feature/registerpost/RegisterPostViewModelTest.kt
@@ -2,7 +2,7 @@ package com.withpeace.withpeace.feature.registerpost
 
 import app.cash.turbine.test
 import com.google.common.truth.Truth.assertThat
-import com.withpeace.withpeace.core.domain.model.LimitedImages
+import com.withpeace.withpeace.core.domain.model.image.LimitedImages
 import com.withpeace.withpeace.core.domain.model.WithPeaceError
 import com.withpeace.withpeace.core.domain.model.post.PostTopic
 import com.withpeace.withpeace.core.domain.model.post.RegisterPost
